[ANDROSDK-1969] move fields and service to network layer
